consolidated files API
