+++ lingo3d/src/api/files/downloadText.ts
@@ -1,4 +1,4 @@
-import downloadBlob from "../../utils/downloadBlob"
+import downloadBlob from "./downloadBlob"
 
 export default (filename: string, data: string) => {
     const blob = new Blob([data], { type: "text/plain" })
