+++ lingo3d/src/api/files/exportVue.ts
@@ -1,7 +1,7 @@
 import { upperFirst, kebabCase } from "@lincode/utils"
-import serialize from "../../api/serializer/serialize"
-import { SceneGraphNode } from "../../api/serializer/types"
-import saveTextFile from "./saveTextFile"
+import serialize from "../serializer/serialize"
+import { SceneGraphNode } from "../serializer/types"
+import downloadText from "./downloadText"
 
 const serializeVue = (nodes: Array<SceneGraphNode>) => {
     let result = ""
@@ -44,5 +44,5 @@ export default async () => {
         </template>
     `, { parser: "vue", plugins: [parser] })
 
-    saveTextFile("App.vue", code)
+    downloadText("App.vue", code)
 }
